Better Config Validation & Better Builds
- Updates the ConfigServer class to use the AJV library to validate the structure of the config JSON.
- Updates the build script to be less verbose by default. Adds a --verbose flag and npm command `npm run buildinfo` so you can still see detailed information.
- Updates the build script to use the Ignore library to handle the `.buildignore` file.
- Updates the build script to run file copy operations in parallel.

+++ build.mjs
@@ -1,23 +1,27 @@
 #!/usr/bin/env node
 
-import fs from 'fs';
-import os from 'os';
-import path from 'path';
-import archiver from 'archiver';
-import winston from 'winston';
+import fs from "fs";
+import os from "os";
+import path from "path";
+import ignore from "ignore";
+import archiver from "archiver";
+import winston from "winston";
+
+const args = process.argv.slice(2);
+const verbose = args.includes("--verbose") || args.includes("-v");
 
 const logLevels = {
     error: 0,
     warn: 1,
-    info: 2,
-    success: 3,
+    success: 2,
+    info: 3,
 };
 
 const logColors = {
-    error: 'red',
-    warn: 'yellow',
-    info: 'grey',
-    success: 'green',
+    error: "red",
+    warn: "yellow",
+    info: "grey",
+    success: "green",
 };
 
 winston.addColors(logColors);
@@ -32,49 +36,68 @@ const logger = winston.createLogger({
     ),
     transports: [
         new winston.transports.Console({
-            level: 'success', // Set the log level to 'success'
+            level: verbose ? "info" : "success",
         }),
     ],
 });
 
 async function main() {
     const currentDir = getCurrentDirectory();
-    const buildIgnorePatterns = await loadBuildIgnoreFile(currentDir);
+    let projectDir;
 
-    const packageJson = await loadPackageJson(currentDir);
-    const projectName = createProjectName(packageJson);
-    logger.log('success', `Project name created: ${projectName}`);
+    try {
+        const buildIgnorePatterns = await loadBuildIgnoreFile(currentDir);
 
-    const distDir = await cleanAndCreateDistDirectory(currentDir);
-    logger.log('success', `Distribution directory cleaned.`);
+        const packageJson = await loadPackageJson(currentDir);
+        const projectName = createProjectName(packageJson);
+        logger.log("success", `Project name created: ${projectName}`);
 
-    const projectDir = await createTemporaryDirectoryWithProjectName(projectName);
-    logger.log('success', `Created temporary working directory: ${projectDir}`);
+        const distDir = await cleanAndCreateDistDirectory(currentDir);
+        logger.log("info", `Distribution directory successfully cleaned.`);
 
-    logger.log('info', 'Beginning copy operation using .buildignore file...');
-    await copyFiles(currentDir, projectDir, buildIgnorePatterns);
-    logger.log('success', 'Files have been successfully copied to temporary directory.');
+        projectDir = await createTemporaryDirectoryWithProjectName(projectName);
+        logger.log("success", `Temporary working directory successfully created.`);
+        logger.log("info", projectDir);
 
-    logger.log('info', 'Beginning folder compression...');
-    const zipFilePath = path.join(path.dirname(projectDir), `${projectName}.zip`);
-    await createZipFile(projectDir, zipFilePath, 'user/mods/' + projectName);
-    logger.log('success', `Created package: ${zipFilePath}`);
+        logger.log("info", "Beginning copy operation using .buildignore file...");
+        await copyFiles(currentDir, projectDir, buildIgnorePatterns);
+        logger.log("success", "Files successfully copied to temporary directory.");
 
-    const zipFileInProjectDir = path.join(projectDir, `${projectName}.zip`);
-    await fs.promises.rename(zipFilePath, zipFileInProjectDir);
-    logger.log('success', `Moved package: ${zipFileInProjectDir}`);
+        logger.log("info", "Beginning folder compression...");
+        const zipFilePath = path.join(path.dirname(projectDir), `${projectName}.zip`);
+        await createZipFile(projectDir, zipFilePath, "user/mods/" + projectName);
+        logger.log("success", "Archive successfully created.");
+        logger.log("info", zipFilePath);
 
-    await fs.promises.rename(projectDir, path.join(distDir));
-    logger.log('success', 'Relocated temporary directory into project distribution directory.');
+        const zipFileInProjectDir = path.join(projectDir, `${projectName}.zip`);
+        await fs.promises.rename(zipFilePath, zipFileInProjectDir);
+        logger.log("success", "Archive successfully moved.");
+        logger.log("info", zipFileInProjectDir);
 
-    await fs.promises.rm(projectDir, { force: true, recursive: true });
-    logger.log('success', 'Cleaned temporary directory.');
+        await fs.promises.rename(projectDir, path.join(distDir));
+        logger.log("success", "Temporary directory successfully moved into project distribution directory.");
 
-    logger.log('success', '------------------------------------');
-    logger.log('success', 'Build script completed successfully!');
-    logger.log('success', "Your mod package has been created in the 'dist' directory:");
-    logger.log('success', `/${path.relative(process.cwd(), path.join(distDir, `${projectName}.zip`))}`);
-    logger.log('success', '------------------------------------');
+        logger.log("success", "------------------------------------");
+        logger.log("success", "Build script completed successfully!");
+        logger.log("success", "Your mod package has been created in the 'dist' directory:");
+        logger.log("success", `/${path.relative(process.cwd(), path.join(distDir, `${projectName}.zip`))}`);
+        logger.log("success", "------------------------------------");
+        if (!verbose) {
+            logger.log("success", "To see a detailed build log, use `npm run buildinfo`.");
+            logger.log("success", "------------------------------------");
+        }
+    } catch (err) {
+        logger.log("error", "An error occurred: " + err);
+    } finally {
+        if (projectDir) {
+            try {
+                await fs.promises.rm(projectDir, { force: true, recursive: true });
+                logger.log("info", "Cleaned temporary directory.");
+            } catch (err) {
+                logger.log("error", "Failed to clean temporary directory: " + err);
+            }
+        }
+    }
 }
 
 function getCurrentDirectory() {
@@ -82,31 +105,32 @@ function getCurrentDirectory() {
 }
 
 async function loadBuildIgnoreFile(currentDir) {
-    const buildIgnorePath = path.join(currentDir, '.buildignore');
+    const buildIgnorePath = path.join(currentDir, ".buildignore");
     try {
-        const fileContent = await fs.promises.readFile(buildIgnorePath, 'utf-8');
-        return fileContent.split('\n').filter(pattern => pattern.trim() !== '');
+        const fileContent = await fs.promises.readFile(buildIgnorePath, "utf-8");
+        const ig = ignore().add(fileContent.split("\n"));
+        return ig;
     } catch (err) {
-        logger.log('warn', 'Failed to read .buildignore file. No files or directories will be ignored.');
-        return [];
+        logger.log("warn", "Failed to read .buildignore file. No files or directories will be ignored.");
+        return ignore(); // Return an ignore instance with no patterns
     }
 }
 
 async function loadPackageJson(currentDir) {
-    const packageJsonPath = path.join(currentDir, 'package.json');
-    const packageJsonContent = await fs.promises.readFile(packageJsonPath, 'utf-8');
+    const packageJsonPath = path.join(currentDir, "package.json");
+    const packageJsonContent = await fs.promises.readFile(packageJsonPath, "utf-8");
     return JSON.parse(packageJsonContent);
 }
 
 function createProjectName(packageJson) {
-    const author = packageJson.author.replace(/\W/g, '').toLowerCase();
-    const name = packageJson.name.replace(/\W/g, '').toLowerCase();
+    const author = packageJson.author.replace(/\W/g, "").toLowerCase();
+    const name = packageJson.name.replace(/\W/g, "").toLowerCase();
     const version = packageJson.version;
     return `${author}-${name}-${version}`;
 }
 
 async function cleanAndCreateDistDirectory(projectDir) {
-    const distPath = path.join(projectDir, 'dist');
+    const distPath = path.join(projectDir, "dist");
     await fs.promises.rm(distPath, { force: true, recursive: true });
 
     await fs.promises.mkdir(distPath);
@@ -114,95 +138,66 @@ async function cleanAndCreateDistDirectory(projectDir) {
 }
 
 async function createTemporaryDirectoryWithProjectName(projectName) {
-    const tempDir = await fs.promises.mkdtemp(path.join(os.tmpdir(), 'spt-mod-build-'));
+    const tempDir = await fs.promises.mkdtemp(path.join(os.tmpdir(), "spt-mod-build-"));
     const projectDir = path.join(tempDir, projectName);
     await fs.promises.mkdir(projectDir);
     return projectDir;
 }
 
-async function copyFiles(srcDir, destDir, buildIgnorePatterns) {
+async function copyFiles(srcDir, destDir, ignoreHandler) {
     try {
-        const entries = await fs.promises.readdir(srcDir, {
-            withFileTypes: true,
-        });
+        const entries = await fs.promises.readdir(srcDir, { withFileTypes: true });
+        const copyOperations = [];
 
         for (const entry of entries) {
             const srcPath = path.join(srcDir, entry.name);
             const destPath = path.join(destDir, entry.name);
             const relativePath = path.relative(process.cwd(), srcPath);
 
-            if (isIgnored(relativePath, buildIgnorePatterns)) {
-                logger.log('info', `Ignored: /${path.relative(process.cwd(), srcPath)}`);
+            if (ignoreHandler.ignores(relativePath)) {
+                logger.log("info", `Ignored: /${path.relative(process.cwd(), srcPath)}`);
                 continue;
             }
 
             if (entry.isDirectory()) {
-                const subentries = await fs.promises.readdir(srcPath, {
-                    withFileTypes: true,
-                });
-                const containsNonIgnoredFiles = subentries.some(subentry => {
-                    const subentryPath = path.join(srcPath, subentry.name);
-                    const relativeSubentryPath = path.relative(process.cwd(), subentryPath);
-                    return !isIgnored(relativeSubentryPath, buildIgnorePatterns);
-                });
-
-                if (containsNonIgnoredFiles) {
-                    await fs.promises.mkdir(destPath);
-                    await copyFiles(srcPath, destPath, buildIgnorePatterns);
-                }
+                await fs.promises.mkdir(destPath);
+                copyOperations.push(copyFiles(srcPath, destPath, ignoreHandler));
             } else {
-                await fs.promises.copyFile(srcPath, destPath);
-                logger.log('info', ` Copied: /${path.relative(process.cwd(), srcPath)}`);
+                copyOperations.push(
+                    fs.promises.copyFile(srcPath, destPath).then(() => {
+                        logger.log("info", `Copied: /${path.relative(process.cwd(), srcPath)}`);
+                    })
+                );
             }
         }
+
+        await Promise.all(copyOperations);
     } catch (err) {
-        logger.log('error', 'Error copying files: ' + err);
+        logger.log("error", "Error copying files: " + err);
     }
-}
-
-function isIgnored(filePath, buildIgnorePatterns) {
-    let shouldBeIgnored = false;
-
-    for (const pattern of buildIgnorePatterns) {
-        const isNegation = pattern.startsWith('!');
-        const regexPattern = isNegation ? pattern.slice(1) : pattern;
-        let regex;
-
-        if (regexPattern.endsWith('/')) {
-            regex = new RegExp(regexPattern.replace('*', '.*') + '.*');
-        } else {
-            regex = new RegExp(regexPattern.replace('*', '.*'));
-        }
-
-        if (regex.test(filePath)) {
-            shouldBeIgnored = isNegation ? false : true;
-        }
-    }
-
-    return shouldBeIgnored;
 }
 
 async function createZipFile(directoryToZip, zipFilePath, containerDirName) {
     return new Promise((resolve, reject) => {
         const output = fs.createWriteStream(zipFilePath);
-        const archive = archiver('zip', {
+        const archive = archiver("zip", {
             zlib: { level: 9 }, // Sets the compression level.
         });
 
-        output.on('close', function () {
-            logger.log('info', 'Archiver has finalized. The output and the file descriptor have closed.');
+        output.on("close", function () {
+            logger.log("info", "Archiver has finalized. The output and the file descriptor have closed.");
             resolve();
         });
 
-        archive.on('warning', function (err) {
-            if (err.code === 'ENOENT') {
-                logger.log('warn', `Archiver issued a warning: ${err.code} - ${err.message}`);
+        archive.on("warning", function (err) {
+            if (err.code === "ENOENT") {
+                logger.log("warn", `Archiver issued a warning: ${err.code} - ${err.message}`);
             } else {
                 reject(err);
             }
         });
 
-        archive.on('error', function (err) {
+        archive.on("error", function (err) {
             reject(err);
         });
 
@@ -213,6 +208,4 @@ async function createZipFile(directoryToZip, zipFilePath, containerDirName) {
 }
 
 // Entry point
-main().catch(err => {
-    logger.log('error', 'An error occurred: ' + err);
-});
+main();
